+++ react/src/components/ui/input-name/InputName.tsx
@@ -1,0 +1,15 @@
+// import React from 'react';
+// import styles from './InputName.module.css';
+
+// const InputName = (prop) => {
+
+//   return (
+//     <>
+//       <h2 className={!hidden ? `${styles.popup} ${styles.popup_hidden}` : `${styles.popup}`}>
+//         Card successfully generated!
+//       </h2>
+//     </>
+//   );
+// };
+
+// export default InputName ;
