+++ react/src/constants/constants.tsx
@@ -1,4 +1,4 @@
-import { IPathes } from '../models/interfaces';
+import { FormState, IPathes } from '../models/interfaces';
 
 const Pathes: IPathes = {
   '/about-us': 'About us',
@@ -7,4 +7,53 @@ const Pathes: IPathes = {
   '/form': 'Form',
 };
 
-export { Pathes };
+const initialState: FormState = {
+  inputName: {
+    error: null,
+  },
+
+  inputSurname: {
+    error: null,
+  },
+
+  inputDate: {
+    value: '',
+    error: null,
+  },
+
+  inputImage: {
+    value: '',
+    error: null,
+  },
+
+  selectOccupation: {
+    value: 'Choose occupation',
+    error: null,
+  },
+
+  inputGender: {
+    value: '',
+    values: {
+      inputGenderFirst: false,
+      inputGenderSecond: false,
+      inputGenderThird: false,
+    },
+    error: null,
+  },
+
+  inputCheckbox: {
+    error: null,
+  },
+
+  cardsList: {
+    error: null,
+    cards: [],
+  },
+
+  popup: {
+    error: null,
+    state: false,
+  },
+};
+
+export { Pathes, initialState };
