+++ react/src/utils/utils.ts
@@ -1,3 +1,5 @@
+import { IInputsGenders } from '../models/interfaces';
+
 const capitalizeFirstLetter = (str: string): string =>
   str.length > 0 ? `${str[0].toUpperCase()}${str.slice(1)}` : '';
 
@@ -64,4 +66,101 @@ const idGenerator = (): string => {
   }${randomNumberGenerator(0, 9)}${randomNumberGenerator(0, 9)}`;
 };
 
-export { capitalizeFirstLetter, idGenerator };
+class Validation {
+  validateCheckbox = (checkbox: HTMLInputElement) => {
+    return checkbox.checked
+      ? { inputCheckbox: { value: checkbox.value, error: null } }
+      : { inputCheckbox: { value: '', error: 'Please accept the agreement!' } };
+  };
+
+  validateName = (name: string) => {
+    const nameWithNoSpaces = name.replaceAll(' ', '');
+    const isNameValid = /^[a-zA-Z]+$/.test(nameWithNoSpaces);
+    return isNameValid &&
+      name.length > 3 &&
+      nameWithNoSpaces[0] === nameWithNoSpaces[0].toUpperCase()
+      ? { inputName: { error: null } }
+      : { inputName: { error: 'Please check your name!' } };
+  };
+
+  validateSurname = (surname: string) => {
+    if (surname.length === 0) {
+      return { inputSurname: { error: null } };
+    }
+    const surnameWithNoSpaces = surname.replaceAll(' ', '');
+    const isSurnameValid = /^[a-zA-Z]+$/.test(surnameWithNoSpaces);
+    return isSurnameValid &&
+      surname.length > 3 &&
+      surnameWithNoSpaces[0] === surnameWithNoSpaces[0].toUpperCase()
+      ? { inputSurname: { error: null } }
+      : { inputSurname: { error: 'Please check your surname!' } };
+  };
+
+  validateDate = (date: string) => {
+    const [year, month, day] = date.split('-');
+    return +day > 0 &&
+      +day <= 31 &&
+      +month > 0 &&
+      +month <= 12 &&
+      +year > 0 &&
+      +year <= new Date().getFullYear()
+      ? { inputDate: { error: null } }
+      : { inputDate: { error: 'Please enter a valid date!' } };
+  };
+
+  validateImage = (fileList: FileList) => {
+    if (fileList.length === 0) {
+      return { inputImage: { error: 'Please attach your image!' } };
+    }
+
+    const [image] = fileList;
+    const { size, type } = image;
+
+    return size < 10 * 1000000 && type.includes('image/')
+      ? { inputImage: { value: URL.createObjectURL(image), error: null } }
+      : { inputImage: { error: 'Please check your image!' } };
+  };
+
+  validateOccupation = (value: string) => {
+    return value === 'Choose occupation'
+      ? { selectOccupation: { value: 'Choose occupation', error: 'Please enter occupation!' } }
+      : { selectOccupation: { value: value, error: null } };
+  };
+
+  validateGender = (genders: HTMLInputElement[]) => {
+    const checkedGender = genders.filter((gender) => gender.checked)[0];
+
+    const inputsGenders: IInputsGenders = {
+      inputGender: {
+        value: '',
+        values: {
+          inputGenderFirst: false,
+          inputGenderSecond: false,
+          inputGenderThird: false,
+        },
+        error: null,
+      },
+    };
+
+    if (!checkedGender) {
+      inputsGenders.inputGender.error = 'Please choose gender!';
+      return inputsGenders;
+    }
+
+    const valueWithRef = {
+      male: 'inputGenderFirst',
+      female: 'inputGenderSecond',
+      'non-binary': 'inputGenderThird',
+    };
+
+    inputsGenders.inputGender.values[
+      valueWithRef[checkedGender.value as keyof typeof valueWithRef]
+    ] = true;
+
+    inputsGenders.inputGender.value = checkedGender.value;
+
+    return inputsGenders;
+  };
+}
+
+export { capitalizeFirstLetter, idGenerator, Validation };
